Allow https:// for urls in \homepage
Up to now, the template always inserted a 'http://' in the url
for the \homepage{} attribute. Therefore, it was not possible
to specify 'https://example.com' as this would have resulted
in 'http://https://example.com'.

The new approach allows to specify the homepage-url in any of the
following ways (which will result in the specified strings)

| url provided to \homepage | hyperlink | displayed url-text |
|---------------------------|-----------|--------------------|
| http://example.com | http://example.com | example.com |
| https://example.com | https://example.com | example.com |
| example.com | http://example.com | example.com |

Note that the third row is equivalent to the current behavior.